+++ lib/patchmaster/code_key.rb
@@ -1,16 +1,16 @@
 module PM
 
-# A CodeKey holds a CodeChunk and remembers what key it is assigned to.
+# A CodeKey holds a block or proc and remembers what key it is assigned to.
 class CodeKey
 
-  attr_accessor :key, :code_chunk
+  attr_accessor :key, :block_or_proc
 
-  def initialize(key, code_chunk)
-    @key, @code_chunk = key, code_chunk
+  def initialize(key, proc = nil, &block)
+    @key, @block_or_proc = key, proc || block
   end
 
-  def run
-    @code_chunk.run
+  def call
+    @block_or_proc.call
   end
 end
 end