+++ test/binding_test.rb
@@ -21,7 +21,7 @@ class BindingTest < Test::Unit::TestCase
   def test_code_binding
     code_bindings = @pm.instance_variable_get("@code_bindings".to_sym)
     code_key = code_bindings[Curses::Key::F3]
-    code_key.run
+    code_key.call
     assert $global_code_key_value == 42
   end