+++ site/changes.md
@@ -9,9 +9,11 @@ title: Changes
 
 # 2.0.0
 
-- New `bank` command. `prog_chg` now only takes one value.
+- New `bank`, `bank_msb`, and `bank_lsb` commands. `prog_chg` now only takes one value.
 
 - Switched to [PortMidi](http://portmedia.sourceforge.net/portmidi/).
+
+- Removed edit and save commands
 
 # 1.2.1